backend/requirements.txt
fastapi==0.110.0
uvicorn==0.27.1
pydantic==2.6.3
sqlalchemy==2.0.28
google-genai==0.2.1
python-multipart==0.0.9
pdfplumber==0.11.0
psycopg2-binary==2.9.9
python-dotenv==1.0.1
pikepdf>=8.0.0
